feat: Creates basic layout of main GUI components
Creates the basic  layout of the app 'azaharTEA'. The basic
layout contains the container tha will be used for the main
components: menubar, linenumbersstrip, editor and footer.

A loader for *.kv files was added in the root directory in order
to be able to separate the .kv files for each GUI component more
easily. The file is named 'kivyloader.py'.